Option menu text fixes
Updated the logic for finding the text_lookup_table to be less convoluted.
Updated the method of updating each language's "warp to hub" string.
Added a special exception for the Chinese language to use the translated version of "travel", because the font being used for the menu in Chinese is not displaying "warp to hub" correctly.